+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ### Changelog
 
-### 0.7.1 (develop)
+### 0.8.0 (develop)
 
 - **Rocketeer can now have specific configurations for stages and connections**
+- **Better handling of multiple connections**
+- **Added facade shortcuts `Rocketeer::execute(Task)` and `Rocketeer::on(connection[s], Task)` to execute commands on the remote servers**
+- Added the `--on` flag to all commands to specify which connections the task should be executed on
 - Added `deploy:flush` to clear Rocketeer's cache of credentials
 
 ### 0.7.0 (stable)